refactor: rename routes and files
